+++ src/evm/abi.rs
@@ -4,7 +4,7 @@
 
 use crate::evm::test_contract_path;
 use crate::mavm::Value;
-use crate::run::{ArbosReceipt, Machine};
+use crate::run::{ArbosReceipt, Machine, load_from_file};
 use crate::uint256::Uint256;
 use ethers_core::utils::keccak256;
 use ethers_signers::Signer;
@@ -1034,544 +1034,6 @@ impl<'a> ArbFunctionTable<'a> {
     }
 }
 
-pub struct _ArbOwner<'a> {
-    pub contract_abi: AbiForContract,
-    wallet: &'a Wallet,
-    my_address: Uint256,
-    debug: bool,
-}
-
-impl<'a> _ArbOwner<'a> {
-    pub fn _new(wallet: &'a Wallet, debug: bool) -> Self {
-        let mut contract_abi =
-            AbiForContract::new_from_file(&builtin_contract_path("ArbOwner")).unwrap();
-        contract_abi.bind_interface_to_address(Uint256::from_u64(107));
-        _ArbOwner {
-            contract_abi,
-            wallet,
-            my_address: Uint256::from_bytes(wallet.address().as_bytes()),
-            debug,
-        }
-    }
-
-    pub fn _give_ownership(
-        &self,
-        machine: &mut Machine,
-        new_owner: Uint256,
-        old_owner: Option<Uint256>,
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            old_owner.unwrap_or(self.my_address.clone()),
-            "giveOwnership",
-            &[ethabi::Token::Address(new_owner.to_h160())],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from(format!(
-                "tx failed: {}",
-                receipts[0]._get_return_code_text()
-            )))
-        }
-    }
-
-    pub fn _add_to_reserve_funds(
-        &self,
-        machine: &mut Machine,
-        amount: Uint256,
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "addToReserveFunds",
-            &[],
-            machine,
-            amount,
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _set_fees_enabled(
-        &self,
-        machine: &mut Machine,
-        enabled: bool,
-        force_owner: bool, // force the message to come from address zero, which is an owner
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            if force_owner {
-                Uint256::zero()
-            } else {
-                self.my_address.clone()
-            },
-            "setFeesEnabled",
-            &[ethabi::Token::Bool(enabled)],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from(format!(
-                "tx failed: {}",
-                receipts[0]._get_return_code_text()
-            )))
-        }
-    }
-
-    pub fn _get_fee_recipients(
-        &self,
-        machine: &mut Machine,
-    ) -> Result<(Uint256, Uint256), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "getFeeRecipient",
-            &[],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if !receipts[0].succeeded() {
-            return Err(ethabi::Error::from("reverted"));
-        }
-
-        let return_vals = ethabi::decode(
-            &[ethabi::ParamType::Address, ethabi::ParamType::Address],
-            &receipts[0].get_return_data(),
-        )?;
-
-        match (&return_vals[0], &return_vals[1]) {
-            (ethabi::Token::Address(addr1), ethabi::Token::Address(addr2)) => Ok((
-                Uint256::from_bytes(addr1.as_bytes()),
-                Uint256::from_bytes(addr2.as_bytes()),
-            )),
-            _ => panic!(),
-        }
-    }
-
-    pub fn _set_fee_recipients(
-        &self,
-        machine: &mut Machine,
-        recipient1: Uint256,
-        recipient2: Uint256,
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "setFeeRecipient",
-            &[
-                ethabi::Token::Address(recipient1.to_h160()),
-                ethabi::Token::Address(recipient2.to_h160()),
-            ],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _set_seconds_per_send(
-        &self,
-        machine: &mut Machine,
-        seconds_per_send: Uint256,
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "setSecondsPerSend",
-            &[ethabi::Token::Uint(seconds_per_send.to_u256())],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _set_gas_accounting_params(
-        &self,
-        machine: &mut Machine,
-        speed_limit: Uint256,
-        gas_pool_max: Uint256,
-        tx_gas_limit: Uint256,
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "setGasAccountingParams",
-            &[
-                ethabi::Token::Uint(speed_limit.to_u256()),
-                ethabi::Token::Uint(gas_pool_max.to_u256()),
-                ethabi::Token::Uint(tx_gas_limit.to_u256()),
-            ],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _change_sequencer(
-        &self,
-        machine: &mut Machine,
-        sequencer_addr: Uint256,
-        delay_seconds: Uint256,
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "changeSequencer",
-            &[
-                ethabi::Token::Address(sequencer_addr.to_h160()),
-                ethabi::Token::Uint(delay_seconds.to_u256()),
-            ],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _start_code_upload(&self, machine: &mut Machine) -> Result<(), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function_compressed(
-            self.my_address.clone(),
-            "startCodeUpload",
-            &[],
-            machine,
-            Uint256::zero(),
-            self.wallet,
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _continue_code_upload(
-        &self,
-        machine: &mut Machine,
-        marshalled_code: Vec<u8>,
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "continueCodeUpload",
-            &[ethabi::Token::Bytes(marshalled_code)],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _get_uploaded_code_hash(
-        &self,
-        machine: &mut Machine,
-    ) -> Result<Uint256, ethabi::Error> {
-        let (receipts, _) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "getUploadedCodeHash",
-            &[],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-        Ok(Uint256::from_bytes(&receipts[receipts.len()-1].get_return_data()))
-    }
-
-    pub fn _finish_code_upload_as_arbos_upgrade(
-        &self,
-        machine: &mut Machine,
-        expected_code_hash: Uint256,
-    ) -> Result<(), ethabi::Error> {
-        let _res = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "finishCodeUploadAsArbosUpgrade",
-            &[ethabi::Token::FixedBytes(expected_code_hash.to_bytes_be())],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-        Ok(())
-    }
-
-    pub fn _finish_code_upload_as_pluggable(
-        &self,
-        machine: &mut Machine,
-        id: Uint256,
-        keep_state: bool,
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "finishCodeUploadAsPluggable",
-            &[
-                ethabi::Token::Uint(id.to_u256()),
-                ethabi::Token::Bool(keep_state),
-            ],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-}
-
-pub struct _ArbGasInfo<'a> {
-    pub contract_abi: AbiForContract,
-    _wallet: &'a Wallet,
-    my_address: Uint256,
-    debug: bool,
-}
-
-impl<'a> _ArbGasInfo<'a> {
-    pub fn _new(wallet: &'a Wallet, debug: bool) -> Self {
-        let mut contract_abi =
-            AbiForContract::new_from_file(&builtin_contract_path("ArbGasInfo")).unwrap();
-        contract_abi.bind_interface_to_address(Uint256::from_u64(108));
-        _ArbGasInfo {
-            contract_abi,
-            _wallet: wallet,
-            my_address: Uint256::from_bytes(wallet.address().as_bytes()),
-            debug,
-        }
-    }
-
-    pub fn _get_prices_in_wei(
-        &self,
-        machine: &mut Machine,
-    ) -> Result<(Uint256, Uint256, Uint256, Uint256, Uint256, Uint256), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "getPricesInWei",
-            &[],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            let return_vals = ethabi::decode(
-                &[
-                    ethabi::ParamType::Uint(256),
-                    ethabi::ParamType::Uint(256),
-                    ethabi::ParamType::Uint(256),
-                    ethabi::ParamType::Uint(256),
-                    ethabi::ParamType::Uint(256),
-                    ethabi::ParamType::Uint(256),
-                ],
-                &receipts[0].get_return_data(),
-            )?;
-
-            match (
-                &return_vals[0],
-                &return_vals[1],
-                &return_vals[2],
-                &return_vals[3],
-                &return_vals[4],
-                &return_vals[5],
-            ) {
-                (
-                    ethabi::Token::Uint(ui0),
-                    ethabi::Token::Uint(ui1),
-                    ethabi::Token::Uint(ui2),
-                    ethabi::Token::Uint(ui3),
-                    ethabi::Token::Uint(ui4),
-                    ethabi::Token::Uint(ui5),
-                ) => Ok((
-                    Uint256::from_u256(&ui0),
-                    Uint256::from_u256(&ui1),
-                    Uint256::from_u256(&ui2),
-                    Uint256::from_u256(&ui3),
-                    Uint256::from_u256(&ui4),
-                    Uint256::from_u256(&ui5),
-                )),
-                _ => panic!(),
-            }
-        } else {
-            Err(ethabi::Error::from(format!(
-                "tx failed: {}",
-                receipts[0]._get_return_code_text()
-            )))
-        }
-    }
-
-    pub fn _get_prices_in_arbgas(
-        &self,
-        machine: &mut Machine,
-    ) -> Result<(Uint256, Uint256, Uint256), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "getPricesInArbGas",
-            &[],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            let return_vals = ethabi::decode(
-                &[
-                    ethabi::ParamType::Uint(256),
-                    ethabi::ParamType::Uint(256),
-                    ethabi::ParamType::Uint(256),
-                ],
-                &receipts[0].get_return_data(),
-            )?;
-
-            match (&return_vals[0], &return_vals[1], &return_vals[2]) {
-                (ethabi::Token::Uint(ui0), ethabi::Token::Uint(ui1), ethabi::Token::Uint(ui2)) => {
-                    Ok((
-                        Uint256::from_u256(&ui0),
-                        Uint256::from_u256(&ui1),
-                        Uint256::from_u256(&ui2),
-                    ))
-                }
-                _ => panic!(),
-            }
-        } else {
-            Err(ethabi::Error::from(format!(
-                "tx failed: {}",
-                receipts[0]._get_return_code_text()
-            )))
-        }
-    }
-
-    pub fn _get_gas_accounting_params(
-        &self,
-        machine: &mut Machine,
-    ) -> Result<(Uint256, Uint256, Uint256), ethabi::Error> {
-        let (receipts, _sends) = self.contract_abi.call_function(
-            self.my_address.clone(),
-            "getGasAccountingParams",
-            &[],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if receipts.len() != 1 {
-            return Err(ethabi::Error::from("wrong number of receipts"));
-        }
-
-        if receipts[0].succeeded() {
-            let return_vals = ethabi::decode(
-                &[
-                    ethabi::ParamType::Uint(256),
-                    ethabi::ParamType::Uint(256),
-                    ethabi::ParamType::Uint(256),
-                ],
-                &receipts[0].get_return_data(),
-            )?;
-
-            match (&return_vals[0], &return_vals[1], &return_vals[2]) {
-                (ethabi::Token::Uint(ui0), ethabi::Token::Uint(ui1), ethabi::Token::Uint(ui2)) => {
-                    Ok((
-                        Uint256::from_u256(&ui0),
-                        Uint256::from_u256(&ui1),
-                        Uint256::from_u256(&ui2),
-                    ))
-                }
-                _ => panic!(),
-            }
-        } else {
-            Err(ethabi::Error::from(format!(
-                "tx failed: {}",
-                receipts[0]._get_return_code_text()
-            )))
-        }
-    }
-}
-
 pub struct ArbosTest {
     pub contract_abi: AbiForContract,
     debug: bool,
@@ -1747,186 +1209,6 @@ impl ArbosTest {
                 "arbosTest.run revert code {}",
                 receipts[0].get_return_code()
             );
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-}
-
-pub struct _ArbAggregator {
-    pub contract_abi: AbiForContract,
-    debug: bool,
-}
-
-impl _ArbAggregator {
-    pub fn _new(debug: bool) -> Self {
-        let mut contract_abi =
-            AbiForContract::new_from_file(&builtin_contract_path("ArbAggregator")).unwrap();
-        contract_abi.bind_interface_to_address(Uint256::from_u64(109));
-        _ArbAggregator {
-            contract_abi,
-            debug,
-        }
-    }
-
-    pub fn _get_preferred_aggregator(
-        &self,
-        machine: &mut Machine,
-        addr: Uint256,
-    ) -> Result<(Uint256, bool), ethabi::Error> {
-        let (receipts, sends) = self.contract_abi.call_function(
-            Uint256::zero(), // send from address zero
-            "getPreferredAggregator",
-            &[ethabi::Token::Address(addr.to_h160())],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if (receipts.len() != 1) || (sends.len() != 0) {
-            Err(ethabi::Error::from("wrong number of receipts or sends"))
-        } else if receipts[0].succeeded() {
-            let return_vals = ethabi::decode(
-                &[ethabi::ParamType::Address, ethabi::ParamType::Bool],
-                &receipts[0].get_return_data(),
-            )?;
-
-            match (&return_vals[0], &return_vals[1]) {
-                (ethabi::Token::Address(r1), ethabi::Token::Bool(r2)) => {
-                    Ok((Uint256::from_bytes(r1.as_bytes()), *r2))
-                }
-                _ => panic!(),
-            }
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _set_preferred_aggregator(
-        &self,
-        machine: &mut Machine,
-        addr: Uint256,
-        sender: Uint256,
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, sends) = self.contract_abi.call_function(
-            sender,
-            "setPreferredAggregator",
-            &[ethabi::Token::Address(addr.to_h160())],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if (receipts.len() != 1) || (sends.len() != 0) {
-            Err(ethabi::Error::from("wrong number of receipts or sends"))
-        } else if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _get_default_aggregator(&self, machine: &mut Machine) -> Result<Uint256, ethabi::Error> {
-        let (receipts, sends) = self.contract_abi.call_function(
-            Uint256::zero(), // send from address zero
-            "getDefaultAggregator",
-            &[],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if (receipts.len() != 1) || (sends.len() != 0) {
-            Err(ethabi::Error::from("wrong number of receipts or sends"))
-        } else if receipts[0].succeeded() {
-            let return_vals = ethabi::decode(
-                &[ethabi::ParamType::Address],
-                &receipts[0].get_return_data(),
-            )?;
-
-            match &return_vals[0] {
-                ethabi::Token::Address(r1) => Ok(Uint256::from_bytes(r1.as_bytes())),
-                _ => panic!(),
-            }
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _set_default_aggregator(
-        &self,
-        machine: &mut Machine,
-        addr: Uint256,
-        sender: Option<Uint256>, // default sender is address zero
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, sends) = self.contract_abi.call_function(
-            if let Some(s) = sender {
-                s
-            } else {
-                Uint256::zero()
-            },
-            "setDefaultAggregator",
-            &[ethabi::Token::Address(addr.to_h160())],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if (receipts.len() != 1) || (sends.len() != 0) {
-            Err(ethabi::Error::from("wrong number of receipts or sends"))
-        } else if receipts[0].succeeded() {
-            Ok(())
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _get_fee_collector(
-        &self,
-        machine: &mut Machine,
-        addr: Uint256,
-    ) -> Result<Uint256, ethabi::Error> {
-        let (receipts, sends) = self.contract_abi.call_function(
-            Uint256::from_u64(10892),
-            "getFeeCollector",
-            &[ethabi::Token::Address(addr.to_h160())],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if (receipts.len() != 1) || (sends.len() != 0) {
-            Err(ethabi::Error::from("wrong number of receipts or sends"))
-        } else if receipts[0].succeeded() {
-            Ok(Uint256::from_bytes(&receipts[0].get_return_data()))
-        } else {
-            Err(ethabi::Error::from("reverted"))
-        }
-    }
-
-    pub fn _set_fee_collector(
-        &self,
-        machine: &mut Machine,
-        agg_addr: Uint256,
-        new_collector: Uint256,
-        sender: Uint256,
-    ) -> Result<(), ethabi::Error> {
-        let (receipts, sends) = self.contract_abi.call_function(
-            sender,
-            "setFeeCollector",
-            &[
-                ethabi::Token::Address(agg_addr.to_h160()),
-                ethabi::Token::Address(new_collector.to_h160()),
-            ],
-            machine,
-            Uint256::zero(),
-            self.debug,
-        )?;
-
-        if (receipts.len() != 1) || (sends.len() != 0) {
-            Err(ethabi::Error::from("wrong number of receipts or sends"))
-        } else if receipts[0].succeeded() {
-            Ok(())
-        } else {
             Err(ethabi::Error::from("reverted"))
         }
     }
