+++ src/evm/mod.rs
@@ -292,267 +292,6 @@ pub fn evm_test_arbsys_direct(log_to: Option<&Path>, debug: bool) -> Result<(), 
     Ok(())
 }
 
-pub fn _evm_test_arbowner(log_to: Option<&Path>, debug: bool) -> Result<(), ethabi::Error> {
-    let mut machine = load_from_file(Path::new("arb_os/arbos.mexe"));
-    machine.start_at_zero();
-
-    let wallet = machine.runtime_env.new_wallet();
-    let my_addr = Uint256::from_bytes(wallet.address().as_bytes());
-
-    let arbowner = _ArbOwner::_new(&wallet, debug);
-
-    arbowner._give_ownership(&mut machine, my_addr, Some(Uint256::zero()))?;
-
-    arbowner._start_code_upload(&mut machine)?;
-
-    let mcode = vec![0x90u8, 1u8, 0u8, 42u8]; // debugprint(42)
-    arbowner._continue_code_upload(&mut machine, mcode)?;
-
-    let expected_code_hash = arbowner._get_uploaded_code_hash(&mut machine)?;
-    arbowner._finish_code_upload_as_arbos_upgrade(
-        &mut machine,
-        expected_code_hash,
-    )?;
-
-    arbowner._set_seconds_per_send(&mut machine, Uint256::from_u64(10))?;
-
-    arbowner._set_gas_accounting_params(
-        &mut machine,
-        Uint256::from_u64(100_000_000),
-        Uint256::from_u64(6_000_000_000),
-        Uint256::from_u64(1_000_000_000),
-    )?;
-
-    if let Some(path) = log_to {
-        machine
-            .runtime_env
-            .recorder
-            .to_file(path, machine.get_total_gas_usage().to_u64().unwrap())
-            .unwrap();
-    }
-
-    Ok(())
-}
-
-pub fn _evm_test_arbgasinfo(log_to: Option<&Path>, debug: bool) -> Result<(), ethabi::Error> {
-    let mut machine = load_from_file(Path::new("arb_os/arbos.mexe"));
-    machine.start_at_zero();
-
-    let wallet = machine.runtime_env.new_wallet();
-    let my_addr = Uint256::from_bytes(wallet.address().as_bytes());
-
-    let arbowner = _ArbOwner::_new(&wallet, debug);
-    let arbgasinfo = _ArbGasInfo::_new(&wallet, debug);
-
-    machine.runtime_env.insert_eth_deposit_message(
-        my_addr.clone(),
-        my_addr.clone(),
-        Uint256::_from_eth(100),
-    );
-    let _ = if debug {
-        machine.debug(None)
-    } else {
-        machine.run(None)
-    };
-
-    let (l2tx, l1calldata, storage, basegas, conggas, totalgas) =
-        arbgasinfo._get_prices_in_wei(&mut machine)?;
-    assert!(l2tx.is_zero());
-    assert!(l1calldata.is_zero());
-    assert!(storage.is_zero());
-    assert!(basegas.is_zero());
-    assert!(conggas.is_zero());
-    assert_eq!(basegas.add(&conggas), totalgas);
-
-    arbowner._set_fees_enabled(&mut machine, true, true)?;
-    machine
-        .runtime_env
-        ._advance_time(Uint256::one(), None, true);
-
-    let (l2tx, l1calldata, storage, basegas, conggas, totalgas) =
-        arbgasinfo._get_prices_in_wei(&mut machine)?;
-    println!(
-        "L2 tx {}, L1 calldata {}, L2 storage {}, base gas {}, congestion gas {}, total gas {}",
-        l2tx, l1calldata, storage, basegas, conggas, totalgas
-    );
-    assert_eq!(l2tx, Uint256::from_u64(642483725000000));
-    assert_eq!(l1calldata, Uint256::from_u64(2778308000000));
-    assert_eq!(storage, Uint256::from_u64(301990000000000));
-    assert_eq!(basegas, Uint256::from_u64(15099500));
-    assert!(conggas.is_zero());
-    assert_eq!(basegas.add(&conggas), totalgas);
-
-    let (l2tx, l1calldata, storage) = arbgasinfo._get_prices_in_arbgas(&mut machine)?;
-    println!(
-        "L2 tx / ag {}, L1 calldata / ag {}, L2 storage / ag {}",
-        l2tx, l1calldata, storage
-    );
-    assert_eq!(l2tx, Uint256::from_u64(42550000));
-    assert_eq!(l1calldata, Uint256::from_u64(184000));
-    assert_eq!(storage, Uint256::from_u64(20000000));
-
-    let (speed_limit, gas_pool_max, tx_gas_limit) =
-        arbgasinfo._get_gas_accounting_params(&mut machine)?;
-    println!(
-        "speed limit {}, pool max {}, tx gas limit {}",
-        speed_limit, gas_pool_max, tx_gas_limit
-    );
-    assert_eq!(speed_limit, Uint256::from_u64(100_000_000));
-    assert_eq!(gas_pool_max, Uint256::from_u64(6_000_000_000));
-    assert_eq!(tx_gas_limit, Uint256::from_u64(1_000_000_000));
-
-    if let Some(path) = log_to {
-        machine
-            .runtime_env
-            .recorder
-            .to_file(path, machine.get_total_gas_usage().to_u64().unwrap())
-            .unwrap();
-    }
-
-    Ok(())
-}
-
-pub fn _evm_test_arbaggregator(log_to: Option<&Path>, debug: bool) -> Result<(), ethabi::Error> {
-    let mut machine = load_from_file(Path::new("arb_os/arbos.mexe"));
-    machine.start_at_zero();
-
-    let wallet = machine.runtime_env.new_wallet();
-    let my_addr = Uint256::from_bytes(wallet.address().as_bytes());
-
-    let arbagg = _ArbAggregator::_new(debug);
-
-    assert_eq!(
-        arbagg._get_fee_collector(&mut machine, my_addr.clone())?,
-        my_addr.clone()
-    );
-
-    let pref_agg = arbagg._get_preferred_aggregator(&mut machine, my_addr.clone())?;
-    assert_eq!(pref_agg, (Uint256::zero(), true));
-
-    let new_pref_agg = Uint256::from_u64(4242);
-    arbagg._set_preferred_aggregator(&mut machine, new_pref_agg.clone(), my_addr.clone())?;
-    let pref_agg = arbagg._get_preferred_aggregator(&mut machine, my_addr.clone())?;
-    assert_eq!(pref_agg, (new_pref_agg, false));
-
-    let def_agg = arbagg._get_default_aggregator(&mut machine)?;
-    assert_eq!(def_agg, Uint256::zero());
-
-    let new_def_agg = Uint256::from_u64(9696);
-    arbagg._set_default_aggregator(&mut machine, new_def_agg.clone(), None)?;
-    let def_agg = arbagg._get_default_aggregator(&mut machine)?;
-    assert_eq!(def_agg, new_def_agg);
-
-    assert!(arbagg
-        ._set_default_aggregator(
-            &mut machine,
-            Uint256::from_u64(12345),
-            Some(my_addr.clone())
-        )
-        .is_err());
-
-    assert_eq!(
-        arbagg._get_fee_collector(&mut machine, my_addr.clone())?,
-        my_addr.clone()
-    );
-
-    let new_collector = Uint256::from_u64(1298031);
-
-    assert!(arbagg
-        ._set_fee_collector(
-            &mut machine,
-            my_addr.clone(),
-            new_collector.clone(),
-            new_collector.clone()
-        )
-        .is_err());
-    assert_eq!(
-        arbagg._get_fee_collector(&mut machine, my_addr.clone())?,
-        my_addr.clone()
-    );
-
-    assert!(arbagg
-        ._set_fee_collector(
-            &mut machine,
-            my_addr.clone(),
-            new_collector.clone(),
-            my_addr.clone()
-        )
-        .is_ok());
-    assert_eq!(
-        arbagg._get_fee_collector(&mut machine, my_addr.clone())?,
-        new_collector.clone()
-    );
-
-    let newer_collector = Uint256::from_u64(589713578913);
-    assert!(arbagg
-        ._set_fee_collector(
-            &mut machine,
-            my_addr.clone(),
-            newer_collector.clone(),
-            my_addr.clone()
-        )
-        .is_err());
-    assert!(arbagg
-        ._set_fee_collector(
-            &mut machine,
-            my_addr.clone(),
-            newer_collector.clone(),
-            new_collector.clone()
-        )
-        .is_ok());
-    assert_eq!(
-        arbagg._get_fee_collector(&mut machine, my_addr.clone())?,
-        newer_collector.clone()
-    );
-
-    if let Some(path) = log_to {
-        machine
-            .runtime_env
-            .recorder
-            .to_file(path, machine.get_total_gas_usage().to_u64().unwrap())
-            .unwrap();
-    }
-
-    Ok(())
-}
-
-pub fn _evm_test_rate_control(log_to: Option<&Path>, debug: bool) -> Result<(), ethabi::Error> {
-    let mut machine = load_from_file(Path::new("arb_os/arbos.mexe"));
-    machine.start_at_zero();
-
-    let wallet = machine.runtime_env.new_wallet();
-    let my_addr = Uint256::from_bytes(wallet.address().as_bytes());
-    let arbowner = _ArbOwner::_new(&wallet, debug);
-
-    arbowner._give_ownership(&mut machine, my_addr, Some(Uint256::zero()))?;
-
-    let const_table = init_constant_table();
-
-    let (r1, r2) = arbowner._get_fee_recipients(&mut machine)?;
-    assert_eq!(&r1, const_table.get("NetFee_defaultRecipient").unwrap());
-    assert_eq!(
-        &r2,
-        const_table.get("CongestionFee_defaultRecipient").unwrap()
-    );
-
-    let new_r1 = r1.add(&Uint256::one());
-    let new_r2 = r2.add(&Uint256::one());
-    arbowner._set_fee_recipients(&mut machine, new_r1.clone(), new_r2.clone())?;
-    let (updated_r1, updated_r2) = arbowner._get_fee_recipients(&mut machine)?;
-    assert_eq!(new_r1, updated_r1);
-    assert_eq!(new_r2, updated_r2);
-
-    if let Some(path) = log_to {
-        machine
-            .runtime_env
-            .recorder
-            .to_file(path, machine.get_total_gas_usage().to_u64().unwrap())
-            .unwrap();
-    }
-
-    Ok(())
-}
-
 pub fn evm_test_function_table_access(
     log_to: Option<&Path>,
     debug: bool,
